segmentation-agent/virtual_env_requirements.txt
bedrock-agentcore-starter-toolkit>=0.1.21
strands-agents
strands-agents-tools
boto3
streamlit
plotly
pandas
